requirements.txt
Django==5.2.8
pypdf==5.1.0
google-generativeai==0.8.3
python-dotenv==1.0.1
